+++ src/Twilio/Rest/Verify/V2/Service/Entity/FactorResource.cs
@@ -60,6 +60,7 @@ namespace Twilio.Rest.Verify.V2.Service.Entity
 
             public static readonly NotificationPlatformsEnum Apn = new NotificationPlatformsEnum("apn");
             public static readonly NotificationPlatformsEnum Fcm = new NotificationPlatformsEnum("fcm");
+            public static readonly NotificationPlatformsEnum None = new NotificationPlatformsEnum("none");
         }
 
         public sealed class TotpAlgorithmsEnum : StringEnum
@@ -419,6 +420,7 @@ namespace Twilio.Rest.Verify.V2.Service.Entity
         /// <param name="configSkew"> The number of past and future time-steps valid at a given time </param>
         /// <param name="configCodeLength"> Number of digits for generated TOTP codes </param>
         /// <param name="configAlg"> The algorithm used to derive the TOTP codes </param>
+        /// <param name="configNotificationPlatform"> The transport technology used to generate the Notification Token </param>
         /// <param name="client"> Client to make requests to Twilio </param>
         /// <returns> A single instance of Factor </returns>
         public static FactorResource Update(string pathServiceSid,
@@ -432,9 +434,10 @@ namespace Twilio.Rest.Verify.V2.Service.Entity
                                             int? configSkew = null,
                                             int? configCodeLength = null,
                                             FactorResource.TotpAlgorithmsEnum configAlg = null,
+                                            string configNotificationPlatform = null,
                                             ITwilioRestClient client = null)
         {
-            var options = new UpdateFactorOptions(pathServiceSid, pathIdentity, pathSid){AuthPayload = authPayload, FriendlyName = friendlyName, ConfigNotificationToken = configNotificationToken, ConfigSdkVersion = configSdkVersion, ConfigTimeStep = configTimeStep, ConfigSkew = configSkew, ConfigCodeLength = configCodeLength, ConfigAlg = configAlg};
+            var options = new UpdateFactorOptions(pathServiceSid, pathIdentity, pathSid){AuthPayload = authPayload, FriendlyName = friendlyName, ConfigNotificationToken = configNotificationToken, ConfigSdkVersion = configSdkVersion, ConfigTimeStep = configTimeStep, ConfigSkew = configSkew, ConfigCodeLength = configCodeLength, ConfigAlg = configAlg, ConfigNotificationPlatform = configNotificationPlatform};
             return Update(options, client);
         }
 
@@ -453,6 +456,7 @@ namespace Twilio.Rest.Verify.V2.Service.Entity
         /// <param name="configSkew"> The number of past and future time-steps valid at a given time </param>
         /// <param name="configCodeLength"> Number of digits for generated TOTP codes </param>
         /// <param name="configAlg"> The algorithm used to derive the TOTP codes </param>
+        /// <param name="configNotificationPlatform"> The transport technology used to generate the Notification Token </param>
         /// <param name="client"> Client to make requests to Twilio </param>
         /// <returns> Task that resolves to A single instance of Factor </returns>
         public static async System.Threading.Tasks.Task<FactorResource> UpdateAsync(string pathServiceSid,
@@ -466,9 +470,10 @@ namespace Twilio.Rest.Verify.V2.Service.Entity
                                                                                     int? configSkew = null,
                                                                                     int? configCodeLength = null,
                                                                                     FactorResource.TotpAlgorithmsEnum configAlg = null,
+                                                                                    string configNotificationPlatform = null,
                                                                                     ITwilioRestClient client = null)
         {
-            var options = new UpdateFactorOptions(pathServiceSid, pathIdentity, pathSid){AuthPayload = authPayload, FriendlyName = friendlyName, ConfigNotificationToken = configNotificationToken, ConfigSdkVersion = configSdkVersion, ConfigTimeStep = configTimeStep, ConfigSkew = configSkew, ConfigCodeLength = configCodeLength, ConfigAlg = configAlg};
+            var options = new UpdateFactorOptions(pathServiceSid, pathIdentity, pathSid){AuthPayload = authPayload, FriendlyName = friendlyName, ConfigNotificationToken = configNotificationToken, ConfigSdkVersion = configSdkVersion, ConfigTimeStep = configTimeStep, ConfigSkew = configSkew, ConfigCodeLength = configCodeLength, ConfigAlg = configAlg, ConfigNotificationPlatform = configNotificationPlatform};
             return await UpdateAsync(options, client);
         }
         #endif